requirements.txt
Django>=4.2.7,<4.3
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.9,<2.10
drf-spectacular>=0.26.5,<0.27
django-cors-headers>=4.3.0,<4.4
Pillow>=10.1.0,<10.2
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0

